Move settings method from Helper to Settings
